requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
google-generativeai
python-multipart
python-dotenv
aiofiles